+++ pkg/timeDate/R/calendar-atoms.R
@@ -20,11 +20,7 @@
 #  atoms,timeDate            Returns date/time atoms from a 'timeDate' object
 #  months,timeDate           Extracts months atom from a 'timeDate' object
 ################################################################################
-
-# ---------------------------------------------------------------------------- #
-# Roxygen Tags
-#' @export
-# ---------------------------------------------------------------------------- #
+ 
 setMethod("atoms", "timeDate", function(x, ...)
 {
     # A function implemented by Diethelm Wuertz
@@ -46,7 +42,26 @@ setMethod("atoms", "timeDate", function(x, ...)
     # Check Class Type:
     if (!inherits(x, "timeDate")) stop("Wrong class type")
 
-    # mdy:
+    ## mdy:
+    ##
+    ## 2023-12-19 GNB:
+    ##
+    ##    if you wonder why this works, see my comments in .formatFinCenterNum(). Indeed, it
+    ##    looks peculiar since the date at GMT may be different from the date at FinCenter!
+    ##
+    ##    To summarize, .formatFinCenterNum() computes (with "gmt2any") a GMT time which will
+    ##    print the same as the time of x in FinCenter. as.POSIXlt() calls
+    ##    .formatFinCenterNum() indirectly, so the same holds for it.
+    ##
+    ## TODO: due to the above, the timeDate methods for as.POSIXlt and as.POSIXct are
+    ##       misleading, they do something somewhat different. Since this seems to be used at
+    ##       a number of places in timeDate, it is not straightforward to fix this (assuming
+    ##       I am correct). One way would be to create a new generic with a slightly
+    ##       different name, maybe not exported, with methods the current methods for
+    ##       as.POSIXlt and as.POSIXct, then rename the calls in the package to use that (for
+    ##       example, the one below). When everything is sorted out the as.POSIXlt and
+    ##       as.POSIXct methods can be fixed.
+    
     X <- as.POSIXlt(x, tz = "GMT")
     Y <- X$year + 1900
     m <- X$mon + 1
@@ -64,10 +79,6 @@ setMethod("atoms", "timeDate", function(x, ...)
 })
 
 
-# ---------------------------------------------------------------------------- #
-# Roxygen Tags
-#' @export
-# ---------------------------------------------------------------------------- #
 setMethod("atoms", "ANY", function(x, ...)
 {
     # A function implemented by Diethelm WUertz
@@ -81,12 +92,20 @@ setMethod("atoms", "ANY", function(x, ...)
 
 ################################################################################
 
-# ---------------------------------------------------------------------------- #
-# Roxygen Tags
-#' @export
-# ---------------------------------------------------------------------------- #
-setMethod("months", "timeDate",
-    function(x, abbreviate=NULL)
+
+## 2023-12-19 GNB: 
+## 
+## TODO: the default base R months() gives a character vector of the names of
+##       the months, (in the current locale, I think). This method returns
+##       integers, which is inconsistent. *Fix this?*
+##       (DONE: Fixed below. Still, eventually remove the numeric option?
+##
+## GNB: for now changing this method to be for missing 'abbreviate' and adding a
+##      method for non-missing. So, the character names can be obtained by
+##      setting it to TRUE or FALSE explicitly.
+## 
+setMethod("months", c("timeDate", "missing"),
+    function(x, abbreviate = NULL)
 {
     # A function implemented by Diethelm Wuertz
     # and improved by Yohan Chalabi
@@ -109,13 +128,86 @@ setMethod("months", "timeDate",
     if (!inherits(x, "timeDate")) stop("Wrong class type")
 
     # Month:
-    ans <- as.POSIXlt(x, tz = "GMT")$mon+1
+    ans <- as.POSIXlt(x, tz = "GMT")$mon + 1
     attr(ans, "control") <- c(FinCenter = finCenter(x))
 
     # Return Value:
     ans
 })
 
+## GNB: see comments above
+setMethod("months", c("timeDate", "ANY"), function(x, abbreviate) {
+    format(x, if(abbreviate) "%b" else "%B")
+})
+
 
 ################################################################################
 
+
+## 2023-12-20
+##    author: GNB:
+##    methods for weekdays() and quarters()
+
+## need this method as the generic weekdays() doesn't have a default for 'abbreviate' and
+## hence can't the S4 methods.
+setMethod("weekdays", c("timeDate", "missing"), function(x, abbreviate) {
+    format(x, "%A")
+})
+
+setMethod("weekdays", c("timeDate", "ANY"), function(x, abbreviate = FALSE) {
+    format(x, if(abbreviate) "%a" else "%A")
+})
+
+setMethod("quarters", "timeDate", function(x) {
+    X <- as.POSIXlt(x, tz = "GMT") 
+    paste0("Q", X$mon %/% 3 + 1) # X$mon is in {0,1,...,11}
+})
+
+
+
+## 2023-12-19
+##    author: GNB:
+##    new function with methods for completion in IDEs
+
+atom_names <-
+    c("year", "month", "day", "hour", "minute", "second",
+      "weekday", "weekday0", "quarter")
+    
+
+setMethod("$", "timeDate", function(x, name) {
+    nam <- as.character(name)
+    ## see GNB's notes for the method for months()
+    X <- as.POSIXlt(x, tz = "GMT")
+    
+    res <- switch(nam,
+                  year     = X$year + 1900,
+                  month    = X$mon + 1 ,
+                  day      = X$mday    ,
+                  hour     = X$hour    ,
+                  minute   = X$min     ,
+                  second   = X$sec     ,
+                  weekday  = {tmp <- X$wday; tmp[tmp == 0] <- 7; tmp},
+                  weekday0 = X$wday, #  Sun - 0
+                  quarter  = paste0("Q", X$mon %/% 3 + 1), # X$mon is in {0,1,...,11}
+                  ## default
+                  stop("unknown (or not implemented) time component requested")
+                  )
+    ## TODO: add names using format(x)?
+    res
+})
+
+
+## TODO: check when .DollarNames() was introduced.
+## DONE: in R-2.10.0 (i.e. forever, since timeDate depends on R >= 3.6.0) 
+##
+##       a method could still be defined but with my findMatches.
+##
+if(getRversion() >= '4.3.0') { ## new function .AtNames(); also: findMatches() is now exported
+    .DollarNames.timeDate <- function(x, pattern = "") {    
+        utils::findMatches(pattern, atom_names)
+    }
+} else { ## simplified matches, .DollarNames available but findMatches not (or not exported)
+    .DollarNames.timeDate <- function(x, pattern = "") {    
+        grep(pattern, atom_names, value = TRUE)
+    }
+}